[*] Add isMoving member + Getter & setter [*] Delete AEntity type
